+++ tests/framework/db/mssql/providers/SchemaProvider.php
@@ -10,7 +10,11 @@ declare(strict_types=1);
 
 namespace yiiunit\framework\db\mssql\providers;
 
+use yii\db\CheckConstraint;
+use yii\db\Constraint;
 use yii\db\DefaultValueConstraint;
+use yii\db\ForeignKeyConstraint;
+use yii\db\IndexConstraint;
 use yiiunit\framework\db\AnyValue;
 
 use function in_array;
@@ -25,6 +29,9 @@ use function in_array;
  */
 final class SchemaProvider extends \yiiunit\base\db\providers\SchemaProvider
 {
+    /**
+     * @phpstan-return array<int, array{array<string, array<string, mixed>>}>
+     */
     public static function expectedColumns(): array
     {
         $result = parent::expectedColumns();
@@ -65,6 +72,16 @@ final class SchemaProvider extends \yiiunit\base\db\providers\SchemaProvider
         return $result;
     }
 
+    /**
+     * @phpstan-return array<
+     *   string,
+     *   array{
+     *     string,
+     *     string,
+     *     Constraint|Constraint[]|CheckConstraint[]|DefaultValueConstraint[]|ForeignKeyConstraint[]|IndexConstraint[]|null,
+     *   },
+     * >
+     */
     public static function constraints(): array
     {
         $result = parent::constraints();
@@ -93,6 +110,9 @@ final class SchemaProvider extends \yiiunit\base\db\providers\SchemaProvider
         return $result;
     }
 
+    /**
+     * @phpstan-return array<int, array{string, string}>
+     */
     public static function quoteTableName(): array
     {
         return [
@@ -107,6 +127,9 @@ final class SchemaProvider extends \yiiunit\base\db\providers\SchemaProvider
         ];
     }
 
+    /**
+     * @phpstan-return array<int, array{string, string}>
+     */
     public static function getTableSchema(): array
     {
         return [
@@ -115,5 +138,29 @@ final class SchemaProvider extends \yiiunit\base\db\providers\SchemaProvider
             ['dbo.profile', 'profile'],
             ['profile', 'profile'],
         ];
+    }
+
+    /**
+     * @phpstan-return array<string, array{string, string}>
+     */
+    public static function unquoteSimpleTableName(): array
+    {
+        return array_merge(parent::unquoteSimpleTableName(), [
+            'quoted' => ['[myTable]', 'myTable'],
+            'embedded closing bracket' => ['[a]]b]', 'a]b'],
+            'multiple embedded closing brackets' => ['[a]]b]]c]', 'a]b]c'],
+        ]);
+    }
+
+    /**
+     * @phpstan-return array<string, array{string, string}>
+     */
+    public static function unquoteSimpleColumnName(): array
+    {
+        return array_merge(parent::unquoteSimpleColumnName(), [
+            'quoted' => ['[myColumn]', 'myColumn'],
+            'embedded closing bracket' => ['[a]]b]', 'a]b'],
+            'multiple embedded closing brackets' => ['[a]]b]]c]', 'a]b]c'],
+        ]);
     }
 }