+++ tests/framework/db/oci/providers/SchemaProvider.php
@@ -11,6 +11,9 @@ declare(strict_types=1);
 namespace yiiunit\framework\db\oci\providers;
 
 use yii\db\CheckConstraint;
+use yii\db\Constraint;
+use yii\db\ForeignKeyConstraint;
+use yii\db\IndexConstraint;
 use yiiunit\framework\db\AnyValue;
 
 /**
@@ -23,6 +26,9 @@ use yiiunit\framework\db\AnyValue;
  */
 final class SchemaProvider extends \yiiunit\base\db\providers\SchemaProvider
 {
+    /**
+     * @phpstan-return array<int, array{array<string, array<string, mixed>>}>
+     */
     public static function expectedColumns(): array
     {
         $result = parent::expectedColumns();
@@ -104,6 +110,16 @@ final class SchemaProvider extends \yiiunit\base\db\providers\SchemaProvider
         return $result;
     }
 
+    /**
+     * @phpstan-return array<
+     *   string,
+     *   array{
+     *     string,
+     *     string,
+     *     Constraint|Constraint[]|CheckConstraint[]|ForeignKeyConstraint[]|IndexConstraint[]|null,
+     * },
+     * >
+     */
     public static function constraints(): array
     {
         $result = parent::constraints();
@@ -191,5 +207,31 @@ final class SchemaProvider extends \yiiunit\base\db\providers\SchemaProvider
         );
 
         return $result;
+    }
+
+    /**
+     * @phpstan-return array<string, array{string, string}>
+     */
+    public static function unquoteSimpleTableName(): array
+    {
+        return [
+            ...parent::unquoteSimpleTableName(),
+            'embedded double quote' => ['"a""b"', 'a"b'],
+            'multiple embedded double quotes' => ['"a""b""c"', 'a"b"c'],
+            'quoted' => ['"myTable"', 'myTable'],
+        ];
+    }
+
+    /**
+     * @phpstan-return array<string, array{string, string}>
+     */
+    public static function unquoteSimpleColumnName(): array
+    {
+        return [
+            ...parent::unquoteSimpleColumnName(),
+            'embedded double quote' => ['"a""b"', 'a"b'],
+            'multiple embedded double quotes' => ['"a""b""c"', 'a"b"c'],
+            'quoted' => ['"myColumn"', 'myColumn'],
+        ];
     }
 }