+++ tests/framework/db/pgsql/SchemaTest.php
@@ -40,6 +40,18 @@ final class SchemaTest extends BaseSchema
         'public',
     ];
 
+    #[DataProviderExternal(SchemaProvider::class, 'unquoteSimpleTableName')]
+    public function testUnquoteSimpleTableName(string $input, string $expected): void
+    {
+        parent::testUnquoteSimpleTableName($input, $expected);
+    }
+
+    #[DataProviderExternal(SchemaProvider::class, 'unquoteSimpleColumnName')]
+    public function testUnquoteSimpleColumnName(string $input, string $expected): void
+    {
+        parent::testUnquoteSimpleColumnName($input, $expected);
+    }
+
     #[DataProviderExternal(SchemaProvider::class, 'expectedColumns')]
     public function testColumnSchema(array $columns): void
     {